Improved line logger output
Also made the synchronisation checking at the end wait for page expansion, which it was not doing. This meant there was a race condition at the end when the job status was being reported. It could often reach the lines of code where the final status was reported before the page fetching jobs had completed.